# UT5/Supuesto práctico proyecto/TejeraSantana_Adoney_supuesto_dsw_ut5/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supuesto UT5 Tejera Santana Adoney",
        "description": "Documentacion de supuesto ut5",
        "version": "11.31"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/adoneytj/tipos_serie": {
            "get": {
                "tags": [
                    "Tipos_serie"
                ],
                "summary": "Listar todos los tipos de serie",
                "description": "Lista todos los tipos de serie, solo admin y entrenadores están autorizados",
                "operationId": "indexTiposSerie",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos_serie"
                ],
                "summary": "Crear un tipo de serie",
                "description": "Crea un tipo de serie, solo admin y entrenadores están autorizados",
                "operationId": "storeTiposSerie",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Serie de netflix"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Serie larga"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "Agún parámetro inválido"
                    }
                }
            }
        },
        "/api/adoneytj/tipos_serie/{id_tipo_serie}": {
            "get": {
                "tags": [
                    "Tipos_serie"
                ],
                "summary": "Obtener un tipo de serie",
                "description": "Obtiene un tipo de serie, solo admin y entrenadores están autorizados",
                "operationId": "showTiposSerie",
                "parameters": [
                    {
                        "name": "id_tipo_serie",
                        "in": "path",
                        "description": "ID del tipo de serie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "No se encuentra el tipo de serie"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos_serie"
                ],
                "summary": "Actualizar un tipo de serie",
                "description": "Actualiza un tipo de serie, solo admin y entrenadores están autorizados",
                "operationId": "updateTiposSerie",
                "parameters": [
                    {
                        "name": "id_tipo_serie",
                        "in": "path",
                        "description": "ID del tipo de serie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Serie de netflix"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Serie larga"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "Agún parámetro inválido o no se encuentra el tipo de serie"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos_serie"
                ],
                "summary": "Borrar un tipo de serie",
                "description": "Borra un tipo de serie, solo admin y entrenadores están autorizados",
                "operationId": "destroyTiposSerie",
                "parameters": [
                    {
                        "name": "id_tipo_serie",
                        "in": "path",
                        "description": "ID del tipo de serie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "No se encuentra el tipo de serie"
                    }
                }
            }
        },
...
    },
...
}